Revert "Actions on Dashboards 1: Import UI from `data-apps-main` (#27756)"
This reverts commit 68dc2e0fa7645eb801c4d36b2976574c4b64796f.
