--- frontend/src/metabase/actions/components/ActionViz/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ActionViz";
