+++ frontend/src/metabase/actions/containers/ActionCreator/FormCreator/utils.ts
@@ -2,13 +2,25 @@ import { t } from "ttag";
 import _ from "underscore";
 
 import { moveElement } from "metabase/core/utils/arrays";
+import { slugify } from "metabase/lib/formatting";
 
 import type {
   ActionFormSettings,
   WritebackAction,
+  FieldSettings,
   FieldSettingsMap,
+  ParameterId,
   Parameter,
 } from "metabase-types/api";
+
+import Field from "metabase-lib/metadata/Field";
+import { TYPE } from "metabase-lib/types/constants";
+
+import { getDefaultFieldSettings } from "../../../utils";
+
+export const getFormTitle = (action: WritebackAction): string => {
+  return action.visualization_settings?.name || action.name || t`Action form`;
+};
 
 export const getSubmitButtonColor = (action: WritebackAction): string => {
   if (action.type === "implicit" && action.kind === "row/delete") {
@@ -33,6 +45,86 @@ export const getSubmitButtonLabel = (action: WritebackAction): string => {
   }
 
   return t`Save`;
+};
+
+export const generateFieldSettingsFromParameters = (
+  params: Parameter[],
+  fields?: Field[],
+) => {
+  const fieldSettings: Record<ParameterId, FieldSettings> = {};
+
+  const fieldMetadataMap = Object.fromEntries(
+    fields?.map(f => [slugify(f.name), f]) ?? [],
+  );
+
+  params.forEach((param, index) => {
+    const field = fieldMetadataMap[param.id]
+      ? new Field(fieldMetadataMap[param.id])
+      : undefined;
+
+    const name = param["display-name"] ?? param.name ?? param.id;
+    const displayName = field?.displayName?.() ?? name;
+
+    fieldSettings[param.id] = getDefaultFieldSettings({
+      id: param.id,
+      name,
+      title: displayName,
+      placeholder: displayName,
+      required: !!param?.required,
+      order: index,
+      description: field?.description ?? "",
+      fieldType: getFieldType(param),
+      inputType: getInputType(param, field),
+      field: field ?? undefined,
+    });
+  });
+  return fieldSettings;
+};
+
+const getFieldType = (param: Parameter): "number" | "string" => {
+  return isNumericParameter(param) ? "number" : "string";
+};
+
+const isNumericParameter = (param: Parameter): boolean =>
+  /integer|float/gi.test(param.type);
+
+export const getInputType = (param: Parameter, field?: Field) => {
+  if (!field) {
+    return isNumericParameter(param) ? "number" : "string";
+  }
+
+  if (field.isFK()) {
+    return field.isNumeric() ? "number" : "string";
+  }
+  if (field.isNumeric()) {
+    return "number";
+  }
+  if (field.isBoolean()) {
+    return "boolean";
+  }
+  if (field.isTime()) {
+    return "time";
+  }
+  if (field.isDate()) {
+    return field.isDateWithoutTime() ? "date" : "datetime";
+  }
+  if (
+    field.semantic_type === TYPE.Description ||
+    field.semantic_type === TYPE.Comment ||
+    field.base_type === TYPE.Structured
+  ) {
+    return "text";
+  }
+  if (
+    field.semantic_type === TYPE.Title ||
+    field.semantic_type === TYPE.Email
+  ) {
+    return "string";
+  }
+  if (field.isCategory() && field.semantic_type !== TYPE.Name) {
+    return "category";
+  }
+  return "string";
 };
 
 export const reorderFields = (
