--- frontend/src/metabase/actions/containers/ActionParametersInputForm/index.ts
@@ -1,3 +1,0 @@
-export { default } from "./ActionParametersInputForm";
-
-export { default as ActionParametersInputModal } from "./ActionParametersInputModal";
