--- frontend/src/metabase/actions/containers/ActionPicker/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ActionPicker";
