+++ frontend/src/metabase/dashboard/actions/data-fetching.js
@@ -108,7 +108,11 @@ const loadingComplete = createThunkAction(
         dispatch(setShowLoadingCompleteFavicon(false));
       }, 3000);
     } else {
-      dispatch(setDocumentTitle(t`Your dashboard is ready`));
+      const dashboard = getDashboardComplete(getState());
+      const message = dashboard.is_app_age
+        ? t`Your page is ready`
+        : t`Your dashboard is ready`;
+      dispatch(setDocumentTitle(message));
       document.addEventListener(
         "visibilitychange",
         () => {
