+++ frontend/src/metabase/dashboard/actions/index.ts
@@ -7,4 +7,3 @@ export * from "./revisions";
 export * from "./save";
 export * from "./sharing";
 export * from "./ui";
-export * from "./actions";
