--- frontend/src/metabase/dashboard/components/AddActionSidebar/index.ts
@@ -1,1 +1,0 @@
-export * from "./AddActionSidebar";
