+++ frontend/src/metabase/dashboard/components/DashCard/DashCardActionButtons/DashCardActionButtons.tsx
@@ -5,11 +5,7 @@ import Icon from "metabase/components/Icon";
 
 import { getVisualizationRaw } from "metabase/visualizations";
 
-import {
-  Dashboard,
-  DashboardOrderedCard,
-  VisualizationSettings,
-} from "metabase-types/api";
+import { Dashboard, VisualizationSettings } from "metabase-types/api";
 import { Series } from "metabase-types/types/Visualization";
 
 import DashCardActionButton from "./DashCardActionButton";
@@ -22,7 +18,6 @@ import { DashCardActionButtonsContainer } from "./DashCardActionButtons.styled";
 interface Props {
   series: Series;
   dashboard: Dashboard;
-  dashcard?: DashboardOrderedCard;
   isLoading: boolean;
   isVirtualDashCard: boolean;
   isPreviewing: boolean;
@@ -37,7 +32,6 @@ interface Props {
 function DashCardActionButtons({
   series,
   dashboard,
-  dashcard,
   isLoading,
   isVirtualDashCard,
   isPreviewing,
@@ -76,7 +70,6 @@ function DashCardActionButtons({
           key="chart-settings-button"
           series={series}
           dashboard={dashboard}
-          dashcard={dashcard}
           onReplaceAllVisualizationSettings={onReplaceAllVisualizationSettings}
         />,
       );
