+++ frontend/src/metabase/dashboard/components/DashCard/DashCardVisualization.tsx
@@ -55,7 +55,6 @@ interface DashCardVisualizationProps {
 
   expectedDuration: number;
   isSlow: CardSlownessStatus;
-  isAction: boolean;
 
   isPreviewing: boolean;
   isEmbed: boolean;
@@ -100,7 +99,6 @@ function DashCardVisualization({
   totalNumGridCols,
   expectedDuration,
   error,
-  isAction,
   headerIcon,
   isSlow,
   isPreviewing,
@@ -203,7 +201,6 @@ function DashCardVisualization({
       errorIcon={error?.icon}
       showTitle
       isDashboard
-      isAction={isAction}
       isSlow={isSlow}
       isFullscreen={isFullscreen}
       isNightMode={isNightMode}
