+++ frontend/src/metabase/dashboard/constants.js
@@ -1,7 +1,5 @@
 export const SIDEBAR_NAME = {
   addQuestion: "addQuestion",
-  addActionButton: "addActionButton",
-  addActionForm: "addActionForm",
   clickBehavior: "clickBehavior",
   editParameter: "editParameter",
   sharing: "sharing",
