+++ frontend/src/metabase/dashboard/containers/DashboardHeader.jsx
@@ -277,23 +277,6 @@ class DashboardHeader extends Component {
         </span>,
       );
 
-      if (canEdit) {
-        buttons.push(
-          <>
-            <DashboardHeaderActionDivider />
-            <Tooltip key="add-action-button" tooltip={t`Add action button`}>
-              <DashboardHeaderButton
-                isActive={activeSidebarName === SIDEBAR_NAME.addActionButton}
-                onClick={() => toggleSidebar(SIDEBAR_NAME.addActionButton)}
-                data-metabase-event={`Dashboard;Add Action Sidebar`}
-              >
-                <Icon name="click" size={18} />
-              </DashboardHeaderButton>
-            </Tooltip>
-          </>,
-        );
-      }
-
       extraButtons.push({
         title: t`Revision history`,
         icon: "history",
