+++ frontend/src/metabase/nav/containers/MainNavbar/types.ts
@@ -17,7 +17,13 @@ export interface MainNavbarDispatchProps {
 export type MainNavbarProps = MainNavbarOwnProps & MainNavbarDispatchProps;
 
 export interface SelectedItem {
-  type: "card" | "collection" | "dashboard" | "non-entity";
+  type:
+    | "card"
+    | "collection"
+    | "dashboard"
+    | "data-app"
+    | "data-app-page"
+    | "non-entity";
   id?: number | string;
   url?: string;
 }
