+++ frontend/src/metabase/parameters/utils/mapping-options.js
@@ -1,5 +1,4 @@
 import { tag_names } from "cljs/metabase.shared.parameters.parameters";
-import { isActionDashCard } from "metabase/actions/utils";
 import Question from "metabase-lib/Question";
 import { ExpressionDimension } from "metabase-lib/Dimension";
 import {
@@ -62,17 +61,6 @@ export function getParameterMappingOptions(
   if (dashcard && card.display === "text") {
     const tagNames = tag_names(dashcard.visualization_settings.text || "");
     return tagNames ? tagNames.map(buildTextTagOption) : [];
-  }
-
-  if (isActionDashCard(dashcard)) {
-    const actionParams = dashcard?.action?.parameters?.map(param => ({
-      icon: "variable",
-      isForeign: false,
-      name: param.id,
-      ...param,
-    }));
-
-    return actionParams || [];
   }
 
   if (!card.dataset_query) {
