+++ frontend/src/metabase/query_builder/components/__mocks__/NativeQueryEditor.jsx
@@ -3,15 +3,12 @@ import React from "react";
 
 import SyncedParametersList from "metabase/parameters/components/SyncedParametersList/SyncedParametersList";
 
-const MockNativeQueryEditor = ({ query, setParameterValue }) => (
-  <div data-testid="mock-native-query-editor">
-    <SyncedParametersList
-      className="mt1"
-      parameters={query.question().parameters()}
-      setParameterValue={setParameterValue}
-      commitImmediately
-    />
-  </div>
+const MockNativeQueryEditor = ({ location, query, setParameterValue }) => (
+  <SyncedParametersList
+    className="mt1"
+    parameters={query.question().parameters()}
+    setParameterValue={setParameterValue}
+    commitImmediately
+  />
 );
-
 export default MockNativeQueryEditor;
