+++ frontend/src/metabase/visualizations/components/ChartSettings.jsx
@@ -252,15 +252,8 @@ class ChartSettings extends Component {
   };
 
   render() {
-    const {
-      className,
-      question,
-      addField,
-      noPreview,
-      dashboard,
-      dashcard,
-      isDashboard,
-    } = this.props;
+    const { className, question, addField, noPreview, dashboard, isDashboard } =
+      this.props;
     const { currentWidget, popoverRef } = this.state;
 
     const settings = this._getSettings();
@@ -405,7 +398,6 @@ class ChartSettings extends Component {
                   isEditing
                   isDashboard
                   dashboard={dashboard}
-                  dashcard={dashcard}
                   isSettings
                   showWarnings
                   onUpdateVisualizationSettings={this.handleChangeSettings}
