+++ frontend/src/metabase/visualizations/register.js
@@ -3,8 +3,6 @@ import {
   setDefaultVisualization,
 } from "metabase/visualizations";
 
-import ActionViz from "metabase/actions/components/ActionViz";
-
 import Scalar from "./visualizations/Scalar";
 import SmartScalar from "./visualizations/SmartScalar";
 import Progress from "./visualizations/Progress";
@@ -43,6 +41,5 @@ export default function () {
   registerVisualization(Funnel);
   registerVisualization(ObjectDetail);
   registerVisualization(PivotTable);
-  registerVisualization(ActionViz);
   setDefaultVisualization(Table);
 }
