+++ frontend/test/register-visualizations.js
@@ -1,13 +1,7 @@
 import "metabase/plugins/builtin";
-
 // We need to mock this *before* registering the visualizations. Otherwise
 // `ChartWithLegend` with already load the real one.
 jest.mock("metabase/components/ExplicitSize");
-
-// We need to mock this *before* registering the visualizations.
-// Otherwise ActionViz loads the NativeQueryEditor (via ActionCreator)
-// and tests fail because ace is not properly mocked
-jest.mock("metabase/query_builder/components/NativeQueryEditor");
 
 import registerVisualizations from "metabase/visualizations/register";
 registerVisualizations();
